+++ NeoForge/src/main/java/de/markusbordihn/adaptiveperformancetweaks/server/ServerEventHandler.java
@@ -165,6 +165,6 @@ public final class ServerEventHandler {
       return;
     }
 
-    PlayerDamageManager.handleLivingDamage(event.getSource(), event.getNewDamage());
+    PlayerDamageManager.handleLivingDamage(event.getSource(), event.getInflictedDamage());
   }
 }